refactor(v1.2.5): split openapi.py into 13 atomic SRP modules + facade
Decomposes openapi.py (500 lines, 6 mixed concerns: config dataclasses,
HTML rendering for 3 UIs, JSON Schema builders, route operation builder,
factory function) into a 13-module package. Public import surface preserved
via __init__.py.

New structure (tachyon_api/openapi/):
  - _info.py             — Contact, License, Info dataclasses
  - _server.py           — Server dataclass
  - _config.py           — OpenAPIConfig + to_openapi_dict()
  - _factory.py          — create_openapi_config() flat-kwargs factory
  - _format_map.py       — _OPENAPI_FORMAT_MAP (datetime/UUID format hints)
  - _safe_json.py        — _safe_json() HTML-safe JSON encoder
  - _struct_schemas.py   — _schema_for_python_type, build_components_for_struct
  - _param_schemas.py    — _scalar_schema, build_param_schema
  - _route_builder.py    — RouteOperationBuilder (builds operation dict per route)
  - _generator.py        — OpenAPIGenerator (spec state + delegation)
  - _swagger_html.py     — SwaggerUIRenderer
  - _redoc_html.py       — RedocRenderer
  - _scalar_html.py      — ScalarRenderer
  - __init__.py          — public re-exports

Public API unchanged: every symbol previously importable from
tachyon_api.openapi remains accessible at the same path.

OpenAPI is cold path — no Cython migration planned. The split is purely
for SRP and testability. Each HTML renderer is now an independent class,
making future UI swaps a single-file change.

Verified: 360/361 tests pass.

Bumps version to 1.2.5.
